AI Developer Assignment/pdf-qa-app/requirements.txt
fastapi
uvicorn
sentence-transformers
PyMuPDF
faiss-cpu
pydantic
streamlit
python-multipart
transformers
requests
streamlit
asyncio
